requirements.txt
Django==1.9
django-inspect==0.4.1
django-realtime==1.1
django-rest-framework-mongoengine==3.3.1
djangorestframework==3.3.3
meld3==1.0.2
mongoengine==0.9.0
pymongo==2.8.1

